services/optimizer/requirements.txt
pandas
python-dotenv
# sqlite3 is part of the Python standard library
